Finished full GUI overhaul
Finished the full overhauling of the project, including a fully functional GUI
